# UF2177actividad3/openapi/uf2177.yaml
openapi: 3.0.1
info:
  title: UF2177
  description: REST para actividad 3 MF0226
  version: 1.0.0
servers:
  - url: 'http://localhost:8080/api/'
    description: Servidor de Spring Boot
tags:
  - name: resena
    description: Reseñas sobre los libros
  - name: libro
    description: Libro almacenado
  - name: autor
    description: Autor del libro
paths:
  '/libros':
    get:
      operationId: getAll, obtener todos los libros
      responses:
        '200':
          description: 'successful operation'
          content:
            application/json:
              schema:
                type: 'array'
                items:
                  $ref: '#/components/schemas/Libro'
  '/resenas':
    post:
      operationId: insertar nueva reseña en un libro
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Resena'
      responses:
        '201':
          description: Creado
        '400':
           description: Petición incorrecta
    '/resenas/{id}':
    get:
      summary: Obtener resenas de un libro por id del libro
      operationId: getById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Devuelve las resenas del libro cuyo id es el que se ha pedido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Resena'
        '404':
          description: No encontrado
components:
  
  schemas:
    Libro:
      type: object
      properties:
        id:
          type: long
          format: int64
        nombre:
          type: string
        autor:
          type: string
    Resena:
      type: object
      properties:
        id:
          type: long
          format: int64
        texto:
          type: string
        libros_id:
          type: long
          format: int64